requirements.txt
pytelegrambotapi==4.14.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
SpeechRecognition==3.10.0
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0

# WEB
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
jinja2==3.1.4
python-dotenv==1.0.0
httpx==0.27.0  # для запросов к API бота